Workaround for Ubuntu/Fedora sudo failure
When running Fedora containers in Github Ubuntu 24.04 runners, 'sudo'
fails due to an authentication error. The issue seems to be due to
/ect/shadow, on Fedora side, not being readable by the root user.

On Fedora, there are extended attributes on this file, which doesn't
seem to be present, when running as a container under Ubuntu 24.04.

This patch adds a workaround that sets '/etc/shadow' to be readable by
the 'root' user, allowing Ansible to run tasks with 'become: true'.